Refactor: create data dir
